+++ PARM.h
@@ -57,18 +57,18 @@ type 'PARM' {
 // Photoshop's Filter Factory has PARM:16
 // Filter Foundry has PARM:16000
 typedef struct {   //structure of FF PARM resource
-	long cbSize;     //size of this structure = 0x2068 (or 0x1C68 for Filter Foundry <1.7)
-	long standalone; //0=original FF, 1=standalone filter
-	long val[8];     //initial values of controls
-	long popDialog;  //1 if need to pop a parameter dialog
-	long unknown1;
-	long unknown2; // Byte 48..52 must be 0 to make obfuscation of FF>=7.0.5 work
-	long unknown3;
-	long map_used[4];   //true if map(n) is used
-	long ctl_used[8];   //true if ctl(n) is used
+	uint32_t cbSize;     //size of this structure = 0x2068 (or 0x1C68 for Filter Foundry <1.7)
+	uint32_t standalone; //0=original FF, 1=standalone filter
+	uint32_t val[8];     //initial values of controls
+	uint32_t popDialog;  //1 if need to pop a parameter dialog
+	uint32_t unknown1;
+	uint32_t unknown2; // Byte 48..52 must be 0 to make obfuscation of FF>=7.0.5 work
+	uint32_t unknown3;
+	uint32_t map_used[4];   //true if map(n) is used
+	uint32_t ctl_used[8];   //true if ctl(n) is used
 	unsigned char category[252];    //Category name
 	// Michael Johannhanwahr's protect flag...
-	long iProtected;            // == 1 means protected
+	uint32_t iProtected;            // == 1 means protected
 	unsigned char title[256];       //Filter title
 	unsigned char copyright[256];   //Copyright info
 	unsigned char author[256];      //Filter author(s)
@@ -79,16 +79,16 @@ typedef struct {   //structure of FF PARM resource
 
 // Premiere's Transition/Filter Factory has PARM:16000
 typedef struct {   //structure of Premiere FF/TF PARM resource
-	long cbSize;    //size of this structure = 0x206C
-	long standalone;  //0=original FF, 1=standalone filter
-	long singleExpression; //1 if "single expression" is checked (member only available in Premiere)
-	long val[8];    //initial values of controls
-	long popDialog; //1 if need to pop a parameter dialog
-	long unknown1;
-	long unknown2;
-	long unknown3;
-	long map_used[4];   //true if map(n) is used
-	long ctl_used[8];   //true if ctl(n) is used
+	uint32_t cbSize;    //size of this structure = 0x206C
+	uint32_t standalone;  //0=original FF, 1=standalone filter
+	uint32_t singleExpression; //1 if "single expression" is checked (member only available in Premiere)
+	uint32_t val[8];    //initial values of controls
+	uint32_t popDialog; //1 if need to pop a parameter dialog
+	uint32_t unknown1;
+	uint32_t unknown2;
+	uint32_t unknown3;
+	uint32_t map_used[4];   //true if map(n) is used
+	uint32_t ctl_used[8];   //true if ctl(n) is used
 	unsigned char title[256]; // in Photoshop Filter Factory: Category
 	unsigned char author[256]; // in Photoshop Filter Factory: Title
 	unsigned char modulename[256]; // in Photoshop Filter Factory: Copyright